Setup as a consumable library.
Add a setup.py. Wrap modules in a directory. Refactor scripts to be
importable and reusable.

One catch: there is no accounting for handling artist disambiguation
apart from via the CLI.
