+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import find_packages, setup
+
+setup(
+    name='playlistmanager',
+    version='0.1',
+    author="Austin Noto-Moniz",
+    author_email="mathfreak65@gmail.com",
+    packages=find_packages(),
+    python_requires='>=3.6'
+)
